cancelTicket

+++ src/main/java/com/example/book_my_show/Models/MovieEntity.java
@@ -13,8 +13,8 @@ import java.util.List;
 
 @Entity
 @Table(name = "movies")
-@Data // no need of getter setter required constructor
-@Builder
+@Data // no need of getter setter required constructor  bt do mention all arg constructor nd no arg constructor
+@Builder // to build entity
 @AllArgsConstructor
 @NoArgsConstructor
 public class MovieEntity {
@@ -22,22 +22,21 @@ public class MovieEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private  int id ;
 
-    @Column(unique = true,nullable = false)
+    @Column(unique = true , nullable = false)
     private String movieName ;
 
-    private double ratings ;
+    private int ratings ;
 
     private  int duration ;
 
     @Enumerated(value = EnumType.STRING)
     private Language language ;
+
    @Enumerated(value = EnumType.STRING)
     private Genre genre ;
+
     //this is parent wrt to shows
     @OneToMany(mappedBy = "movieEntity",cascade = CascadeType.ALL)
     private List<ShowEntity> showEntityList = new ArrayList<>();
 
-
-
-
-}
+  }
